Only check for at-zero during roll-under
The double-counting problem only exists when rolling under from 0 to 99.
Going from 99 to 0 (aka 100) is what gets us to the 0 value in the first
place, which is counted in the first check (`if next == 0`).

Since all conditions are doing the same action, I could flatten these
down into one big `if` condition, but I want to keep space for the text
print outs so I won't be combining it.